feat: expose encrypt/decrypt utilities
